Added logic which removes afk players, those who did not do anything in roles of attacker/defender for 20 seconds.
TODO:
Please fix EndCurrentRound method, so that when attacking player wants to finish, we start the timer like in surrendering and kill timer for afk, when other players add more attacking cards, we stop the timer, and start afk timer.

Also would be nice to show finishing round timer on the client.